agents/requirements.txt
bedrock-agentcore
strands-agents>=0.2.0
strands-agents-tools>=0.2.10
